requirements.txt
google-generativeai
python-dotenv
flask
flask-cors
jsonify
request
webbrowser
threading
time